+++ plugins/GSdx/CMakeLists.txt
@@ -213,7 +213,6 @@ set(GSdxHeaders
 if(Windows)
     LIST(APPEND GSdxSources
         Renderers/DX11/GSDevice11.cpp
-        Renderers/DXCommon/GSDeviceDX.cpp
         Window/GSDialog.cpp
         Renderers/DX11/GSRendererDX11.cpp
         Renderers/DX11/GSTexture11.cpp
@@ -230,7 +229,6 @@ if(Windows)
 
     LIST(APPEND GSdxHeaders
         Renderers/DX11/GSDevice11.h
-        Renderers/DXCommon/GSDeviceDX.h
         Renderers/DX11/GSRendererDX11.h
         Renderers/DX11/GSTexture11.h
         Renderers/DX11/GSTextureCache11.h
